refactor : RemoveTodo 분리
