+++ src/components/todo/TodoItem.tsx
@@ -1,8 +1,6 @@
-/* eslint-disable no-alert */
-import { FaSpinner, FaTrash } from 'react-icons/fa';
+import React from 'react';
 import { TodoListType } from 'type/todo';
-import React, { useCallback, useState } from 'react';
-import { deleteTodo } from 'api/todo';
+import { RemoveTodo } from './RemoveTodo';
 
 type TodoItemProps = {
   id: string;
@@ -11,33 +9,10 @@ type TodoItemProps = {
 };
 
 const TodoItem = ({ id, title, setTodos }: TodoItemProps) => {
-  const [isLoading, setIsLoading] = useState<boolean>(false);
-
-  const handleRemoveTodo = useCallback(async () => {
-    try {
-      setIsLoading(true);
-      await deleteTodo(id);
-
-      setTodos(prev => prev.filter(item => item.id !== id));
-    } catch (error) {
-      alert('Something went wrong.');
-    } finally {
-      setIsLoading(false);
-    }
-  }, [id, setTodos, setIsLoading]);
-
   return (
     <li className="item">
       <span>{title}</span>
-      <div className="item-option">
-        {!isLoading ? (
-          <button onClick={() => handleRemoveTodo()}>
-            <FaTrash className="btn-trash" />
-          </button>
-        ) : (
-          <FaSpinner className="spinner" />
-        )}
-      </div>
+      <RemoveTodo id={id} setTodos={setTodos} />
     </li>
   );
 };
